move code to `runtime` file
